+++ airbyte-ci/connectors/pipelines/pipelines/actions/secrets.py
@@ -9,7 +9,7 @@ import datetime
 from typing import TYPE_CHECKING
 
 from dagger import Secret
-from pipelines.actions.environments import python
+from pipelines.actions.environments.common import tools
 from pipelines.utils import get_file_contents, get_secret_host_variable
 
 if TYPE_CHECKING:
@@ -44,7 +44,7 @@ async def download(context: ConnectorContext, gcp_gsm_env_variable_name: str = "
     """
     gsm_secret = get_secret_host_variable(context.dagger_client, gcp_gsm_env_variable_name)
     secrets_path = f"/{context.connector.code_directory}/secrets"
-    ci_credentials = await python.with_ci_credentials(context, gsm_secret)
+    ci_credentials = await tools.with_ci_credentials(context, gsm_secret)
     with_downloaded_secrets = (
         ci_credentials.with_exec(["mkdir", "-p", secrets_path])
         .with_env_variable(
@@ -81,7 +81,7 @@ async def upload(context: ConnectorContext, gcp_gsm_env_variable_name: str = "GC
     gsm_secret = get_secret_host_variable(context.dagger_client, gcp_gsm_env_variable_name)
     secrets_path = f"/{context.connector.code_directory}/secrets"
 
-    ci_credentials = await python.with_ci_credentials(context, gsm_secret)
+    ci_credentials = await tools.with_ci_credentials(context, gsm_secret)
 
     return await ci_credentials.with_directory(secrets_path, context.updated_secrets_dir).with_exec(
         ["ci_credentials", context.connector.technical_name, "update-secrets"]
